+++ src/cli.ts
@@ -1,61 +1,58 @@
 #! /usr/bin/env node
 
 /* NOTE: calling this file entropy.ts helps commander parse process.argv */
-import { Command, Option } from 'commander'
+import { Command } from 'commander'
 
-import { EntropyTuiOptions } from './types'
-import { coreVersion, loadEntropy, tuiEndpointOption, versionOption } from './common/utils-cli'
 import * as config from './config'
-
-import launchTui from './tui'
-import { entropyAccountCommand } from './account/command'
-import { entropyTransferCommand } from './transfer/command'
-import { entropySignCommand } from './sign/command'
-import { entropyBalanceCommand } from './balance/command'
-import { entropyProgramCommand } from './program/command'
 import { print } from './common/utils'
 
-const program = new Command()
+import { entropyTuiCommand as tui, /* tuiAction */ } from './tui'
+import { entropyAccountCommand as account } from './account/command'
+import { entropyTransferCommand as transfer } from './transfer/command'
+import { entropySignCommand as sign } from './sign/command'
+import { entropyBalanceCommand as balance } from './balance/command'
+import { entropyProgramCommand as program } from './program/command'
+
+const packageVersion = 'v' + require('../package.json').version
+const coreVersion = process.env.ENTROPY_CORE_VERSION.split('-')[1]
+
+const cli = new Command()
 
 /* no command */
-program
+cli
   .name('entropy')
-  .description('CLI interface for interacting with entropy.xyz. Running this binary without any commands or arguments starts a text-based interface.')
-  .addOption(
-    new Option(
-      '-d, --dev',
-      'Runs entropy in a developer mode uses the dev endpoint as the main endpoint and allows for faucet option to be available in the main menu'
-    )
-      .env('DEV_MODE')
-      .hideHelp()
-  )
-  .addOption(tuiEndpointOption())
-  .addOption(versionOption())
-  .addOption(coreVersion())
-  .addCommand(entropyBalanceCommand())
-  .addCommand(entropyAccountCommand())
-  .addCommand(entropyTransferCommand())
-  .addCommand(entropySignCommand())
-  .addCommand(entropyProgramCommand())
+  .description('CLI interface for interacting with entropy.xyz.')
 
-  .action(async (opts: EntropyTuiOptions) => {
-    const { account, tuiEndpoint, version, coreVersion } = opts
-    const entropy = account
-      ? await loadEntropy(account, tuiEndpoint)
-      : undefined
-    if (version) {
-      print(`v${version}`)
+  .addCommand(tui())
+  .addCommand(account())
+  .addCommand(sign())
+  .addCommand(balance())
+  .addCommand(transfer())
+  .addCommand(program())
+
+  .option('-v, --version', 'Displays the current running version of Entropy CLI')
+  .option('-cv, --core-version', 'Displays the current running version of the Entropy Protocol')
+  .action(opts => {
+    if (opts.version) {
+      print(packageVersion)
       process.exit(0)
-    } else if (coreVersion) {
+    }
+    if (opts.coreVersion) {
       print(coreVersion)
       process.exit(0)
     }
-    // NOTE: on initial startup you have no account
-    launchTui(entropy, opts)
+
+    // print entropy help and exit
+    cli.help()
+
+    // tuiAction(opts)
+    // NOTE: this doesn't quite work, because -a, -e are not defined as options
+    // and if we do put them in here it gets a bit confusing
   })
+
+  // set up config file, run migrations
   .hook('preAction', async () => {
-    // set up config file, run migrations
     return config.init()
   })
 
-program.parseAsync()
+cli.parseAsync()